requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2024.2.0
ftfy==6.2.3
huggingface-hub==0.24.7
idna==3.8
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
open_clip_torch==2.26.1
packaging==24.1
pillow==10.2.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==70.0.0
sympy==1.12
tabulate==0.9.0
timm==1.0.9
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.66.5
typing_extensions==4.9.0
urllib3==2.2.3
wcwidth==0.2.13
